src/app/shared/file-uploader-module/file-uploader-with-params/components/parametrized-files-table/parametrized-files-table.component.spec.ts
/* tslint:disable:no-unused-variable */
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { By } from '@angular/platform-browser';
import { DebugElement } from '@angular/core';

import { ParametrizedFilesTableComponent } from './parametrized-files-table.component';

describe('ParametrizedFilesTableComponent', () => {
  let component: ParametrizedFilesTableComponent;
  let fixture: ComponentFixture<ParametrizedFilesTableComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ ParametrizedFilesTableComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ParametrizedFilesTableComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
});
